重构 cron 子命令，删除 update、factor 子命令，指标视图 sql 移动到目录保存
